src/test/java/be/jforce/nosql/dao/cassandra/assignment/Commuter.java
package be.jforce.nosql.dao.cassandra.assignment;

public class Commuter {
    private final String email;

    private String name;

    private String homeAddress;

    private String workAddress;

    public Commuter(String email) {
        this.email = email;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getEmail() {
        return email;
    }

    public String getName() {
        return name;
    }

    public String getHomeAddress() {
        return homeAddress;
    }

    public String getWorkAddress() {
        return workAddress;
    }

    public void setHomeAddress(String homeAddress) {
        this.homeAddress = homeAddress;
    }

    public void setWorkAddress(String workAddress) {
        this.workAddress = workAddress;
    }
}
